att frequency

+++ Codificacoes/AMICodification.cs
@@ -25,7 +25,7 @@ internal class AMICodification : ILineCodification {
         return newSeq;
     }
 
-    public int GetFrequency() {
+    public double GetFrequency() {
         return 1;
     }
