+++ Codificacoes/PseudoTernaryCodification.cs
@@ -26,7 +26,7 @@ internal class PseudoTernaryCodification : ILineCodification {
         return newSeq;
     }
 
-    public int GetFrequency() {
+    public double GetFrequency() {
         return 1;
     }
